Replace exam example with image example
